mise à jour du côté elec

+++ src/Entity/MesureResistance.php
@@ -17,10 +17,10 @@ class MesureResistance
     private ?int $id = null;
 
     #[ORM\Column]
-    private ?int $temp_ambiante = null;
+    private ?float $temp_ambiante = null;
 
     #[ORM\Column]
-    private ?int $temp_tolerie = null;
+    private ?float $temp_tolerie = null;
 
     #[ORM\Column]
     private ?float $hygrometrie = null;
@@ -47,24 +47,24 @@ class MesureResistance
         return $this->id;
     }
 
-    public function getTempAmbiante(): ?int
+    public function getTempAmbiante(): ?float
     {
         return $this->temp_ambiante;
     }
 
-    public function setTempAmbiante(?int $temp_ambiante): self
+    public function setTempAmbiante(?float $temp_ambiante): self
     {
         $this->temp_ambiante = $temp_ambiante;
 
         return $this;
     }
 
-    public function getTempTolerie(): ?int
+    public function getTempTolerie(): ?float
     {
         return $this->temp_tolerie;
     }
 
-    public function setTempTolerie(?int $temp_tolerie): self
+    public function setTempTolerie(?float $temp_tolerie): self
     {
         $this->temp_tolerie = $temp_tolerie;
