+++ src/Entity/MesureVibratoire.php
@@ -29,46 +29,46 @@ class MesureVibratoire
     private ?string $commentaire = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $n10 = null;
+    private ?float $n10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $n25 = null;
+    private ?float $n25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $a10 = null;
+    private ?float $a10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $a25 = null;
+    private ?float $a25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $b10 = null;
+    private ?float $b10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $b25 = null;
+    private ?float $b25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $c10 = null;
+    private ?float $c10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $c25 = null;
+    private ?float $c25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $d10 = null;
+    private ?float $d10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $d25 = null;
+    private ?float $d25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $e10 = null;
+    private ?float $e10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $e25 = null;
+    private ?float $e25 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $f10 = null;
+    private ?float $f10 = null;
 
     #[ORM\Column(nullable: true)]
-    private ?int $f25 = null;
+    private ?float $f25 = null;
 
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $obervation = null;
@@ -144,168 +144,168 @@ class MesureVibratoire
         return $this;
     }
 
-    public function getN10(): ?int
+    public function getN10(): ?float
     {
         return $this->n10;
     }
 
-    public function setN10(?int $n10): self
+    public function setN10(?float $n10): self
     {
         $this->n10 = $n10;
 
         return $this;
     }
 
-    public function getN25(): ?int
+    public function getN25(): ?float
     {
         return $this->n25;
     }
 
-    public function setN25(?int $n25): self
+    public function setN25(?float $n25): self
     {
         $this->n25 = $n25;
 
         return $this;
     }
 
-    public function getA10(): ?int
+    public function getA10(): ?float
     {
         return $this->a10;
     }
 
-    public function setA10(?int $a10): self
+    public function setA10(?float $a10): self
     {
         $this->a10 = $a10;
 
         return $this;
     }
 
-    public function getA25(): ?int
+    public function getA25(): ?float
     {
         return $this->a25;
     }
 
-    public function setA25(?int $a25): self
+    public function setA25(?float $a25): self
     {
         $this->a25 = $a25;
 
         return $this;
     }
 
-    public function getB10(): ?int
+    public function getB10(): ?float
     {
         return $this->b10;
     }
 
-    public function setB10(?int $b10): self
+    public function setB10(?float $b10): self
     {
         $this->b10 = $b10;
 
         return $this;
     }
 
-    public function getB25(): ?int
+    public function getB25(): ?float
     {
         return $this->b25;
     }
 
-    public function setB25(?int $b25): self
+    public function setB25(?float $b25): self
     {
         $this->b25 = $b25;
 
         return $this;
     }
 
-    public function getC10(): ?int
+    public function getC10(): ?float
     {
         return $this->c10;
     }
 
-    public function setC10(?int $c10): self
+    public function setC10(?float $c10): self
     {
         $this->c10 = $c10;
 
         return $this;
     }
 
-    public function getC25(): ?int
+    public function getC25(): ?float
     {
         return $this->c25;
     }
 
-    public function setC25(?int $c25): self
+    public function setC25(?float $c25): self
     {
         $this->c25 = $c25;
 
         return $this;
     }
 
-    public function getD10(): ?int
+    public function getD10(): ?float
     {
         return $this->d10;
     }
 
-    public function setD10(?int $d10): self
+    public function setD10(?float $d10): self
     {
         $this->d10 = $d10;
 
         return $this;
     }
 
-    public function getD25(): ?int
+    public function getD25(): ?float
     {
         return $this->d25;
     }
 
-    public function setD25(?int $d25): self
+    public function setD25(?float $d25): self
     {
         $this->d25 = $d25;
 
         return $this;
     }
 
-    public function getE10(): ?int
+    public function getE10(): ?float
     {
         return $this->e10;
     }
 
-    public function setE10(?int $e10): self
+    public function setE10(?float $e10): self
     {
         $this->e10 = $e10;
 
         return $this;
     }
 
-    public function getE25(): ?int
+    public function getE25(): ?float
     {
         return $this->e25;
     }
 
-    public function setE25(?int $e25): self
+    public function setE25(?float $e25): self
     {
         $this->e25 = $e25;
 
         return $this;
     }
 
-    public function getF10(): ?int
+    public function getF10(): ?float
     {
         return $this->f10;
     }
 
-    public function setF10(?int $f10): self
+    public function setF10(?float $f10): self
     {
         $this->f10 = $f10;
 
         return $this;
     }
 
-    public function getF25(): ?int
+    public function getF25(): ?float
     {
         return $this->f25;
     }
 
-    public function setF25(?int $f25): self
+    public function setF25(?float $f25): self
     {
         $this->f25 = $f25;
